Put the new TimeCountdown class in core_utils and use it instead of TimeCounter
