+++ backend_py/libs/core_utils/pyproject.toml
@@ -18,3 +18,9 @@ build-backend = "poetry.core.masonry.api"
 [tool.black]
 line-length = 120
 
+[tool.pylint]
+disable = ["missing-docstring", "line-too-long", "logging-fstring-interpolation", "duplicate-code"]
+
+[tool.mypy]
+ignore_missing_imports = true
+disallow_untyped_defs = true
